fix: move wandb to loggers
